+++ CKeyframeSequence.h
@@ -11,17 +11,17 @@ public:
 	void setPosition(glm::vec3 position);
 	void setRotation(glm::quat rotation);
 	void setUVMatrix(glm::mat3x3 uv_matrix, int level);
-	void setBeginFrame(int frame); //begin frame, reletive to the beginning of the sequence collection
+	void setBeginFrame(float frame); //begin frame, reletive to the beginning of the sequence collection
 
 	glm::quat getRotation();
 	glm::vec3 getPosition();
 	glm::mat3x3 getUVMatrix(int level);
-	int getBeginFrame();
+	float getBeginFrame();
 private:
 	glm::quat rotation;
 	glm::vec3 position;
 	glm::mat3x3 uv_matrices[MAX_MESH_TEXTURES];
-	int start_frame;
+	float start_frame;
 };
 
 #endif //_CKEYFRAMESEQUENCE_H
